requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
python-dotenv>=1.0
Pillow>=10.0
